requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
python-dotenv==1.0.0
supabase==2.3.0
redis==4.6.0
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0
python-jose==3.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest==7.4.2
black==23.7.0
flake8==6.1.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
cryptography==41.0.0
langdetect==1.0.9
deep-translator==1.11.4
replicate==0.22.0
opencv-python==4.9.0.80
mediapipe==0.10.9
rembg==2.0.60
ffmpeg-python==0.2.0
stripe==7.11.0